bounds2d, ScratchPad, PrettyLine (anti-aliased lines)
Added some cool classes to enable me to draw soft lines.  So far really
enjoying them, you take 2 verts, a thickness (for each vert) and a
color, and it draws a really cool anti-aliased line.  The anti-aliasing
can be homed in for speed, but generally is set for 4x anti-aliasing.